requirements.txt
python-dotenv
flask
flask-cors
langchain
langchain-openai
langchain-community
langchain-core
openai
aiohttp
requests
httpx
chromadb
pydantic
certifi
numpy
gunicorn